chore(repo): tidy husky hooks and remove stray file; refactor ETag utility and add unit tests; docs: quick Maps setup + OpenAPI examples
