+++ .husky/pre-commit
@@ -1,4 +1,2 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
-
-npx lint-staged
+#!/usr/bin/env sh
+npx --no-install lint-staged
