+++ package.json
@@ -15,7 +15,6 @@
     "test:coverage": "jest --coverage",
     "test:e2e": "jest --config jest.e2e.config.js",
     "validate": "npm run lint && npm run format:check && npm run test",
-    "prepare": "npm run validate",
     "postinstall": "husky install",
     "openapi:lint": "spectral lint design/to-be/openapi.yaml",
     "smoke": "node scripts/smoke.js",
